hms-core/requirements.txt
docutils==0.14
lockfile==0.12.2
pkg-resources==0.0.0
pymongo==3.6.1
PyYAML==3.13
